refactor(pages/CollagePage): 상태를 모두 Page에서 관리하도록 변경
